image_app/requirements.txt
requests
aws-lambda-powertools[parser, validation, tracer]
pillow